Added slides and handouts (PDFs only)
"Long" version of NT slides has been removed, since it's only for
presentations anyway (while "Short" was for reading, which is the
intended purpose of this repository).

This is up to date with what I've written so far. ToC slides should stay
frozen, but the notes still have some work left to do, mainly in
applications. Expect an update by the day of the dev discussion.