Updated ToC notes: string-search
Got started properly on the KMP section, and rewrote part of the
preceding DFA string-search section. Mostly in order to have it ease
into the KMP section a little more easily, by keeping the notation
consistent and emphasizing that KMP is just a clever way to eliminate
most of the work.

I'm not 1000% sure about the prefix-suffix deal at the moment. Will
figure it out next time I work on this.